# contracts/saas/openapi.yaml
openapi: 3.1.0
...
info:
  title: Dockling hosted API contract
  version: 0.4.1
  description: Versioned SaaS design. Hosted Phases 1–4 are implemented for synthetic
    local acceptance; billing, public deployment and future aliases remain planned.
    Uploads/workbook apply use bounded raw bytes. AI uses separate consent and conservative
    reservations; production teardown recovery remains a deployment gate.
servers:
- url: /api/v1
security:
- SessionCookie: []
paths:
...
  /jobs:
    get:
      operationId: get_jobs
      summary: GET /jobs
      x-phase: 1
      x-feature-ids:
      - F02
      x-access-roles:
      - owner
      - editor
      - viewer
      x-workspace-scoped: true
      responses:
        '200':
          description: Successful authorized result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobList'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '413':
          $ref: '#/components/responses/Error413'
        '422':
          $ref: '#/components/responses/Error422'
        '429':
          $ref: '#/components/responses/Error429'
        '503':
          $ref: '#/components/responses/Error503'
      parameters:
      - $ref: '#/components/parameters/QueryLimit'
      - $ref: '#/components/parameters/QueryCursor'
      - name: search
        in: query
        required: false
        schema:
          type: string
          maxLength: 100
    post:
      operationId: post_jobs
      summary: POST /jobs
      x-phase: 1
      x-feature-ids:
      - F01
      x-access-roles:
      - owner
      - editor
      x-workspace-scoped: true
      responses:
        '201':
          description: Successful authorized result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '413':
          $ref: '#/components/responses/Error413'
        '422':
          $ref: '#/components/responses/Error422'
        '429':
          $ref: '#/components/responses/Error429'
        '503':
          $ref: '#/components/responses/Error503'
      parameters:
      - $ref: '#/components/parameters/HeaderXCsrfToken'
      - $ref: '#/components/parameters/HeaderIdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JobOptions'
...
  /settings:
    get:
      summary: Settings
      operationId: hosted_get__settings
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
                title: Response Settings Api V1 Settings Get
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '422':
          $ref: '#/components/responses/Error422'
        '429':
          $ref: '#/components/responses/Error429'
        '503':
          $ref: '#/components/responses/Error503'
      x-phase: 4
      x-implementation: 'implemented: SaaS Phase 4; single API/worker local topology,
        synthetic acceptance only'
      x-workspace-scoped: true
      x-access-roles:
      - owner
      - editor
      - viewer
      x-feature-ids:
      - F32
    put:
      summary: Save Preferences
      operationId: hosted_put__settings
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HostedPreferences'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
                title: Response Save Preferences Api V1 Settings Put
        '422':
          $ref: '#/components/responses/Error422'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '429':
          $ref: '#/components/responses/Error429'
        '503':
          $ref: '#/components/responses/Error503'
      x-phase: 4
      x-implementation: 'implemented: SaaS Phase 4; single API/worker local topology,
        synthetic acceptance only'
      x-workspace-scoped: true
      x-access-roles:
      - owner
      x-feature-ids:
      - F32
      parameters:
      - $ref: '#/components/parameters/HeaderXCsrfToken'
      - $ref: '#/components/parameters/HeaderIdempotencyKey'
  /support-grants:
    get:
      operationId: get_support_grants
      summary: GET /support-grants
      x-phase: 4
      x-feature-ids:
      - N04
      x-access-roles:
      - owner
      x-workspace-scoped: true
      responses:
        '200':
          description: Successful authorized result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SupportGrantList'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '413':
          $ref: '#/components/responses/Error413'
        '422':
          $ref: '#/components/responses/Error422'
        '429':
          $ref: '#/components/responses/Error429'
        '503':
          $ref: '#/components/responses/Error503'
      parameters:
      - $ref: '#/components/parameters/QueryLimit'
      - $ref: '#/components/parameters/QueryCursor'
    post:
      operationId: post_support_grants
      summary: POST /support-grants
      x-phase: 4
      x-feature-ids:
      - N04
      x-access-roles:
      - owner
      x-workspace-scoped: true
      responses:
        '201':
          description: Successful authorized result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SupportGrant'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '413':
          $ref: '#/components/responses/Error413'
        '422':
          $ref: '#/components/responses/Error422'
        '429':
          $ref: '#/components/responses/Error429'
        '503':
          $ref: '#/components/responses/Error503'
      parameters:
      - $ref: '#/components/parameters/HeaderXCsrfToken'
      - $ref: '#/components/parameters/HeaderIdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SupportGrantCreate'
...
  /admin/profiles/{profile}/tests:
    post:
      summary: Profile Test
      operationId: hosted_post__admin_profiles_profile_tests
      parameters:
      - name: profile
        in: path
        required: true
        schema:
          type: string
          title: Identifier
      - $ref: '#/components/parameters/HeaderXCsrfToken'
      - $ref: '#/components/parameters/HeaderIdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HostedProfileTest'
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Profile Test Api V1 Admin Profiles  Identifier  Tests
                  Post
        '422':
          $ref: '#/components/responses/Error422'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '429':
          $ref: '#/components/responses/Error429'
        '503':
          $ref: '#/components/responses/Error503'
      x-phase: 4
      x-implementation: 'implemented: SaaS Phase 4; single API/worker local topology,
        synthetic acceptance only'
      x-workspace-scoped: true
      x-access-roles:
      - platform_admin
      x-feature-ids:
      - F29
      - F32
...
  /connections/{connection}/test:
    post:
      summary: Models
      operationId: hosted_post__connections_connection_test
      parameters:
      - name: connection
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Identifier
      - $ref: '#/components/parameters/HeaderXCsrfToken'
      - $ref: '#/components/parameters/HeaderIdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HostedRevisionCommand'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HostedConnectionRead'
        '422':
          $ref: '#/components/responses/Error422'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '429':
          $ref: '#/components/responses/Error429'
        '503':
          $ref: '#/components/responses/Error503'
      x-phase: 4
      x-implementation: 'implemented: SaaS Phase 4; single API/worker local topology,
        synthetic acceptance only'
      x-workspace-scoped: true
      x-access-roles:
      - owner
      x-feature-ids:
      - F24
      - F25
      - F26
      - F27
...
  /admin/selftest:
    post:
      summary: Selftest
      operationId: hosted_post__admin_selftest
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
                title: Response Selftest Api V1 Admin Selftest Post
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '422':
          $ref: '#/components/responses/Error422'
        '429':
          $ref: '#/components/responses/Error429'
        '503':
          $ref: '#/components/responses/Error503'
      x-phase: 4
      x-implementation: 'implemented: SaaS Phase 4; single API/worker local topology,
        synthetic acceptance only'
      x-workspace-scoped: true
      x-access-roles:
      - platform_admin
      x-feature-ids:
      - F29
      - F31
      parameters:
      - $ref: '#/components/parameters/HeaderXCsrfToken'
      - $ref: '#/components/parameters/HeaderIdempotencyKey'
...
  /admin/support-grants:
    get:
      summary: Assigned
      operationId: hosted_get__admin_support_grants
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
                title: Response Assigned Api V1 Admin Support Grants Get
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '422':
          $ref: '#/components/responses/Error422'
        '429':
          $ref: '#/components/responses/Error429'
        '503':
          $ref: '#/components/responses/Error503'
      x-phase: 4
      x-implementation: 'implemented: SaaS Phase 4; single API/worker local topology,
        synthetic acceptance only'
      x-workspace-scoped: true
      x-access-roles:
      - platform_admin
      x-feature-ids:
      - F29
      - F32
...
components:
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: dockling_session
      description: Opaque server session; Secure, HttpOnly, SameSite. Not an OIDC
        access token or browser role.
    PaymentSignature:
      type: apiKey
      in: header
      name: Payment-Signature
      description: Placeholder for the selected provider signature scheme; validate
        raw provider bytes, timestamp and replay policy before processing.
  schemas:
...
    JobOptions:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
        currency:
          type: string
          maxLength: 3
          pattern: ^[A-Z]{3}$
        date_order:
          type: string
          enum:
          - auto
          - DMY
          - MDY
          - YMD
        output_date_format:
          type: string
          enum:
          - ISO
          - MM/DD/YYYY
          - DD/MM/YYYY
        outputs:
          type: array
          items: &id013
            type: string
            enum:
            - excel
            - csv
            - qb_csv3
            - qb_csv4
            - xero_csv
            - ofx
          maxItems: 6
          minItems: 1
          uniqueItems: true
        merge: *id003
        categorize: *id003
        combine: *id003
        engine:
          type: string
          enum:
          - auto
          - text
          - docling
        profile_id:
          type:
          - string
          - 'null'
          maxLength: 80
          pattern: ^[a-z0-9_-]+$
          description: Choose an active published profile from intake-options; custom
            profiles require synthetic test evidence before activation.
        pages:
          anyOf:
          - type: string
            maxLength: 250
          - type: 'null'
        ai_selection:
          anyOf:
          - $ref: '#/components/schemas/AiSelection'
          - type: 'null'
      required:
      - name
      - currency
      - date_order
      - outputs
      additionalProperties: false
...
    JobList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/JobSummary'
          maxItems: 100
        next_cursor:
          anyOf:
          - type: string
            maxLength: 500
          - type: 'null'
      required:
      - items
      - next_cursor
      additionalProperties: false
    Job:
      type: object
      properties:
        id: *id001
        name: *id006
        status: *id007
        revision: *id004
        currency: *id008
        created_at: *id005
        deletion_state: *id009
        options:
          $ref: '#/components/schemas/JobOptions'
        content_digest:
          anyOf:
          - &id010
            type: string
            pattern: ^[0-9a-f]{64}$
          - type: 'null'
        export_revision:
          anyOf:
          - *id004
          - type: 'null'
        source_checked: *id003
        ai_source_checked: *id003
      required:
      - id
      - name
      - status
      - revision
      - currency
      - created_at
      - deletion_state
      - options
      - content_digest
      - export_revision
      - source_checked
      - ai_source_checked
      additionalProperties: false
...
    SupportGrant:
      type: object
      properties:
        id: *id001
        version: *id004
        job_ids:
          type: array
          items: *id001
          maxItems: 20
        administrator_id: *id001
        expires_at: *id005
        data_classes:
          type: array
          items:
            type: string
            enum:
            - source
            - rows
            - checks
            - outputs
          maxItems: 4
        actions:
          type: array
          items:
            type: string
            enum:
            - limit_override
            - operator_metadata
            - export_exception
            - private_scaffold
          maxItems: 4
        active: *id003
      required:
      - id
      - version
      - job_ids
      - administrator_id
      - expires_at
      - data_classes
      - actions
      - active
      additionalProperties: false
    SupportGrantCreate:
      type: object
      properties:
        job_ids:
          type: array
          items: *id001
          maxItems: 20
          minItems: 1
        administrator_id: *id001
        expires_at: *id005
        data_classes:
          type: array
          items:
            type: string
            enum:
            - source
            - rows
            - checks
            - outputs
          maxItems: 4
        actions:
          type: array
          items:
            type: string
            enum:
            - limit_override
            - operator_metadata
            - export_exception
            - private_scaffold
          maxItems: 4
      required:
      - job_ids
      - administrator_id
      - expires_at
      - data_classes
      additionalProperties: false
    SupportGrantList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/SupportGrant'
          maxItems: 100
        next_cursor:
          anyOf:
          - type: string
            maxLength: 500
          - type: 'null'
      required:
      - items
      - next_cursor
      additionalProperties: false
...
    HostedPreferences:
      properties:
        expected_version:
          type: integer
          minimum: 0.0
          title: Expected Version
        currency:
          type: string
          pattern: ^[A-Z]{3}$
          title: Currency
        date_order:
          type: string
          enum:
          - auto
          - DMY
          - MDY
          - YMD
          title: Date Order
        output_date_format:
          type: string
          enum:
          - ISO
          - MM/DD/YYYY
          - DD/MM/YYYY
          title: Output Date Format
        outputs:
          items:
            type: string
            enum:
            - excel
            - csv
            - qb_csv3
            - qb_csv4
            - xero_csv
            - ofx
          type: array
          maxItems: 6
          minItems: 1
          title: Outputs
      additionalProperties: false
      type: object
      required:
      - expected_version
      - currency
      - date_order
      - output_date_format
      - outputs
      title: Preferences
    HostedProfileTest:
      properties:
        profile:
          additionalProperties: true
          type: object
          title: Profile
      additionalProperties: false
      type: object
      required:
      - profile
      title: ProfileTest
...
  responses:
    Error400:
      description: Safe structured error without private input echo
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Error401:
      description: Safe structured error without private input echo
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Error403:
      description: Safe structured error without private input echo
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Error404:
      description: Nonrevealing missing or unauthorized private resource
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Error409:
      description: Revision/idempotency/state conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Error413:
      description: Safe structured error without private input echo
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Error422:
      description: Safe structured error without private input echo
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Error429:
      description: Safe structured error without private input echo
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Error503:
      description: Dependency unavailable; no demo fallback
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  parameters:
    HeaderXCsrfToken:
      name: X-CSRF-Token
      in: header
      required: true
      schema:
        type: string
        maxLength: 150
    QueryLimit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 25
    QueryCursor:
      name: cursor
      in: query
      schema:
        type: string
        maxLength: 500
...
    HeaderIdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        maxLength: 128
        minLength: 16
      description: Scoped durable replay key; changed payload under the same key is
        a 409 conflict.
...